# openapi.yaml
openapi: 3.1.3
info:
  title: Chirpy API
  version: 1.0.0
  description: API specification for the Chirpy backend service.
servers:
  - url: http://localhost:42069
    description: Local dev
paths:
...
  /api/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      description: Refreshes the access token using a valid refresh token.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  description: The refresh token
            example:
              refresh_token: "some-refresh-token"
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: "new-access-token"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    polkaApiKey:
      type: apiKey
      in: header
      name: Authorization
      description: 'Authorization: ApiKey <api-key>'
...
security:
  - bearerAuth: []